+++ samples/hello/hello_computed.rb
@@ -16,25 +16,55 @@ class HelloComputed
   def launch
     root {
       title 'Hello, Computed!'
-      frame {
-        label {text 'First Name: '}
+      
+      frame {        
+        grid column: 0, row: 0, padx: 5, pady: 5
+        
+        label {
+          grid column: 0, row: 0, sticky: 'w'
+          text 'First Name: '
+        }
         entry {
+          grid column: 1, row: 0
+          width 15
           text bind(@contact, :first_name)
         }
-        label {text 'Last Name: '}
+        
+        label {
+          grid column: 0, row: 1, sticky: 'w'
+          text 'Last Name: '
+        }
         entry {
+          grid column: 1, row: 1
+          width 15
           text bind(@contact, :last_name)
         }
-        label {text 'Year of Birth: '}
+        
+        label {
+          grid column: 0, row: 2, sticky: 'w'
+          text 'Year of Birth: '
+        }
         entry {
+          grid column: 1, row: 2
+          width 15
           text bind(@contact, :year_of_birth)
         }
-        label {text 'Name: '}
+        
         label {
+          grid column: 0, row: 3, sticky: 'w'
+          text 'Name: '
+        }
+        label {
+          grid column: 1, row: 3, sticky: 'w'
           text bind(@contact, :name, computed_by: [:first_name, :last_name])
         }
-        label {text 'Age: '}
+        
         label {
+          grid column: 0, row: 4, sticky: 'w'
+          text 'Age: '
+        }
+        label {
+          grid column: 1, row: 4, sticky: 'w'
           text bind(@contact, :age, on_write: :to_i, computed_by: [:year_of_birth])
         }
       }